// src/main/webapp/app/shared/model/enumerations/requisition-status.model.ts
export const enum RequisitionStatus {
  PENDING = 'PENDING',
  LM_APPROVED = 'LM_APPROVED',
  HOD_APPROVED = 'HOD_APPROVED',
  CTO_APPROVED = 'CTO_APPROVED',
  HOHR_VETTED = 'HOHR_VETTED',
  CEO_APPROVED = 'CEO_APPROVED',
  OPEN = 'OPEN',
  CLOSED = 'CLOSED',
  PARTIALLY_CLOSED = 'PARTIALLY_CLOSED',
  NOT_APPROVED = 'NOT_APPROVED',
  IN_PROGRESS = 'IN_PROGRESS',
}